requirements.txt
tensorflow-gpu==1.2
certifi==2015.04.28
itchat==1.3.9
jieba==0.38
numpy==1.13.0
